CSRF Fixes

+++ modules/admin/delcours.php
@@ -79,6 +79,10 @@ if (isset($search) && ($search=="yes")) {
 }
 // Delete course
 if (isset($_GET['delete']) && isset($_GET['c']))  {
+    if ($token != $_SESSION['tok']){
+        echo 'Request error!';
+        die;
+    }
 	db_query("DROP DATABASE `".mysql_real_escape_string($_GET['c'])."`");
         mysql_select_db($mysqlMainDb);
         $code = quote($_GET['c']);
@@ -101,7 +105,7 @@ else {
     <td><br />".$langCourseDelConfirm2." <em>".htmlspecialchars($_GET['c'])."</em>;<br /><br /><i>".$langNoticeDel."</i><br /><br /></td>
   </tr>";
 	$tool_content .= "  <tr>
-    <td><ul><li><a href=\"".$_SERVER['PHP_SELF']."?c=".htmlspecialchars($_GET['c'])."&amp;delete=yes".$searchurl."\"><b>$langYes</b></a><br />&nbsp;</li>
+    <td><ul><li><a href=\"".$_SERVER['PHP_SELF']."?c=".htmlspecialchars($_GET['c']).."&amp;token=".$_SESSION['tok']."&amp;delete=yes".$searchurl."\"><b>$langYes</b></a><br />&nbsp;</li>
   <li><a href=\"listcours.php?c=".htmlspecialchars($_GET['c'])."".$searchurl."\"><b>$langNo</b></a></li></ul></td>
   </tr>";
 	$tool_content .= "</tbody></table><br />";
